HTTPDebugProcessor is related with http, not log, then move it into utils_http
