+++ src/motuclient/motu_utils/utils_log.py
@@ -30,15 +30,9 @@
 import logging
 import sys
 if sys.version_info > (3, 0):
-    from urllib.request import BaseHandler
     from urllib.parse import unquote
 else:
-    from urllib2 import BaseHandler
     from urllib2 import unquote
-
-
-# trace level
-TRACE_LEVEL = 1
 
 
 def log_url(log, message, url, level=logging.DEBUG):
@@ -60,34 +54,3 @@ def log_url(log, message, url, level=logging.DEBUG):
             if(len(param) < 2):
                 param.append('')
             log.log(level, ' . %s = %s', unquote(param[0]), unquote(param[1]))
-
-
-class HTTPDebugProcessor(BaseHandler):
-    """ Track HTTP requests and responses with this custom handler.
-    """
-
-    def __init__(self, log, log_level=TRACE_LEVEL):
-        self.log_level = log_level
-        self.log = log
-
-    def http_request(self, request):
-        host, full_url = request.host, request.get_full_url()
-        url_path = full_url[full_url.find(host) + len(host):]
-        log_url(self.log, "Requesting: ", request.get_full_url(), TRACE_LEVEL)
-        self.log.log(self.log_level, "%s %s" %
-                     (request.get_method(), url_path))
-
-        for header in request.header_items():
-            self.log.log(self.log_level, " . %s: %s" % header[:])
-
-        return request
-
-    def http_response(self, request, response):
-        code, msg, headers = response.code, response.msg, response.info().items()
-        self.log.log(self.log_level, "Response:")
-        self.log.log(self.log_level, " HTTP/1.x %s %s" % (code, msg))
-
-        for key, value in headers:
-            self.log.log(self.log_level, " . %s: %s" % (key, value))
-
-        return response
